requirements.txt
# OrphaMind — Python dependencies
# Install with: pip install -r requirements.txt

# ── API Framework ────────────────────────────────────────
fastapi>=0.115.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9

# ── Pydantic ─────────────────────────────────────────────
pydantic>=2.5.0
pydantic-settings>=2.1.0

# ── Google Gemini ─────────────────────────────────────────
google-generativeai>=0.8.0

# ── Vector Database ───────────────────────────────────────
chromadb>=0.6.0

# ── OCR ───────────────────────────────────────────────────
pytesseract>=0.3.10
Pillow>=10.0.0

# ── Data / NLP ───────────────────────────────────────────
numpy>=1.24.0
pandas>=2.0.0
requests>=2.31.0

# ── File Parsing ─────────────────────────────────────────
PyPDF2>=3.0.0
python-pptx>=0.6.21
lxml>=4.9.0

# ── Utilities ────────────────────────────────────────────
python-dotenv>=1.0.0

# ── Optional: local LLM fallback ─────────────────────────
# llama-cpp-python>=0.2.0   # uncomment to enable local LLM mode
